refactor: introduce aws_generic agent tag
