+++ Dockerfile
@@ -62,6 +62,11 @@ FROM base AS generic
 CMD . $VENV_DIR/bin/activate \
     && gunicorn --bind :$PORT --workers $GUNICORN_WORKERS --threads $GUNICORN_THREADS --timeout $GUNICORN_TIMEOUT apollo.interfaces.generic.main:app
 
+FROM base AS aws_generic
+
+CMD . $VENV_DIR/bin/activate \
+    && gunicorn --bind :$PORT --workers $GUNICORN_WORKERS --threads $GUNICORN_THREADS --timeout $GUNICORN_TIMEOUT apollo.interfaces.aws_generic.main:app
+
 FROM base AS cloudrun
 
 COPY requirements-cloudrun.txt ./
